+++ src/app/blog/containers/blog/blog.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ContentfulService } from '../../../services/contentful.service';
 
 @Component({
   selector: 'app-blog',
@@ -7,9 +8,10 @@ import { Component, OnInit } from '@angular/core';
 })
 export class BlogComponent implements OnInit {
 
-  constructor() { }
+  constructor(private contentful: ContentfulService) { }
 
   ngOnInit() {
+    this.contentful.logBlogPosts();
   }
 
 }